requirements.txt
Flask==1.1.2
Flask-SocketIO==5.0.1
python-engineio==4.0.0
python-socketio==5.0.4
Pillow==8.1.0
opencv-python==4.2.0.32
numpy==1.19.5